fix: new not found page
