+++ client/app/[locale]/not-found.tsx
@@ -1,11 +1,23 @@
-import Link from "next/link"
+"use client"
 
-export default function NotFound() {
+import { useRouter } from "next/navigation"
+
+import { Button } from "@/components/ui/button"
+
+const NotFound = () => {
+  const router = useRouter()
+  const handleClick = () => {
+    router.back()
+  }
   return (
-    <div>
-      <h2>Not Found</h2>
-      <p>Could not find requested resource</p>
-      <Link href="/">Return Home</Link>
+    <div className="flex h-[500px] flex-col items-center justify-center gap-4">
+      <span className="text-5xl">🥲😢</span>
+      <h6>{"404 Not Found"}</h6>
+      <Button type="button" onClick={handleClick}>
+        返回
+      </Button>
     </div>
   )
 }
+
+export default NotFound
